Automate with Github Actions
Using a CI enables us to quickly check if we messed up somewhere, though in this case it'll be limited to code mistakes only. We can also use CI to automatically push releases to Github, removing a step that would take away our precious screen staring time.

Ideally we also want it to update a comment in a Pull Request, so that binaries are available directly in the Pull Request, always pointing at the latest changes. But that's a bit difficult